Split Brunch actual mode route
