+++ apps/petrinaut-website/src/main/app/brunch-demo/brunch-demo-app.tsx
@@ -1,5 +1,4 @@
 import { useSentryFeedbackAction } from "../sentry-feedback-button";
-
 import { BrunchActualModeRoute } from "./brunch-actual-mode-route";
 
 export const BrunchDemoApp = () => {
